+++ h5features/entry.py
@@ -92,16 +92,20 @@ class Entry(object):
 
     def _create_dataset(self, group, chunk_size):
         """Create an empty dataset in a group."""
-        shape = (0,) if self.dim == 1 else (0, self.dim)
-        maxshape = (None,) if self.dim == 1 else (None, self.dim)
-
         # if dtype is a variable str, guess representative size is 20 bytes
         per_chunk = (nb_per_chunk(20, self.dim, chunk_size)
                      if self.dtype == np.dtype('O') else
                      nb_per_chunk(np.dtype(self.dtype).itemsize,
                                   self.dim, chunk_size))
+
+        shape = (0,) if self.dim == 1 else (0, self.dim)
+        maxshape = (None,) if self.dim == 1 else (None, self.dim)
         chunks = (per_chunk,) if self.dim == 1 else (per_chunk, self.dim)
+        # shape = (0, self.dim)
+        # maxshape = (None, self.dim)
+        # chunks = (per_chunk, self.dim)
 
         # raise if per_chunk >= 4 Gb, this is requested by h5py
-        group.create_dataset(self.name, shape, dtype=self.dtype,
-                             chunks=chunks, maxshape=maxshape)
+        group.create_dataset(
+            self.name, shape, dtype=self.dtype,
+            chunks=chunks, maxshape=maxshape)