+++ h5features/labels.py
@@ -22,64 +22,84 @@ from .entry import Entry
 
 
 class Labels(Entry):
-    """This class manages labels related operations for h5features files."""
+    """This class manages labels related operations for h5features files
 
-    def __init__(self, data, check=True):
-        dim = self.parse_dim(data, check)
-        super(Labels, self).__init__('labels', data, dim, np.float64, check)
+    :param labels: Each element of the list contains the labels of
+        an h5features item. Empty list are not accepted. For all t
+        in labels, we must have t.ndim to be either 1 or 2.
+
+        * 1D arrays contain the center labelstamps of each frame of the
+          related item.
+
+        * 2D arrays contain the begin and end labelstamps of each
+          items's frame, thus having t.ndim == 2 and t.shape[1] == 2.
+
+    :type labels: list of numpy arrays
+
+    :param bool check: If True, raise on errors
+
+    :raise IOError: if the time format is not 1 or 2, or if labels
+       arrays have different dimensions.
+
+    :return: The parsed labels dimension is either 1 or 2 for 1D
+        or 2D labels arrays respectively.
+
+    """
+    def __init__(self, labels, check=True):
+        if check:
+            self.check(labels)
+        dim = self.parse_dim(labels)
+
+        super(Labels, self).__init__('labels', labels, dim, np.float64, check)
 
     @staticmethod
-    def parse_dim(labels, check=True):
-        """Return the labels vectors dimension.
+    def check(labels):
+        """Raise IOError if labels are not correct
 
-        :param labels: Each element of the list contains the labels of
-            an h5features item. Empty list are not accepted. For all t
-            in labels, we must have t.ndim to be either 1 or 2.
-
-            * 1D arrays contain the center labelstamps of each frame of the
-              related item.
-
-            * 2D arrays contain the begin and end labelstamps of each
-              items's frame, thus having t.ndim == 2 and t.shape[1] == 2.
-
-        :type labels: list of numpy arrays
-
-        :param bool check: If True, raise on errors
-
-        :raise IOError: if the time format is not 1 or 2, or if labels
-            arrays have different dimensions.
-
-        :return: The parsed labels dimension is either 1 or 2 for 1D
-            or 2D labels arrays respectively.
+        `labels` must be a list of sorted numpy arrays of equal
+             dimensions (must be 1D or 2D). In the case of 2D labels,
+             the second axis must have the same shape for all labels.
 
         """
-        if check:
-            # paranoid check because was buggy
-            if not isinstance(labels, list):
-                raise IOError('labels are not in a list')
-            if not len(labels):
-                raise IOError('The labels list is empty')
-            if not all([isinstance(l, np.ndarray) for l in labels]):
-                raise IOError('All labels must be numpy arrays')
+        # type checking
+        if not isinstance(labels, list):
+            raise IOError('labels are not in a list')
+        if not len(labels):
+            raise IOError('the labels list is empty')
+        if not all([isinstance(l, np.ndarray) for l in labels]):
+            raise IOError('all labels must be numpy arrays')
 
-            ndim = labels[0].ndim
-            if ndim not in [1, 2]:
-                raise IOError('Labels dimension must be 1 or 2')
-            if not all([l.ndim == ndim for l in labels]):
-                raise IOError('All labels dimensions must be equal')
-            if ndim == 2:
-                shape1 = labels[0].shape[1]
-                if not all([l.shape[1] == shape1 for l in labels]):
-                    raise IOError('All labels must have same shape on 2nd dim')
+        # dimension checking
+        ndim = labels[0].ndim
+        if ndim not in [1, 2]:
+            raise IOError('labels dimension must be 1 or 2')
+        if not all([l.ndim == ndim for l in labels]):
+            raise IOError('all labels dimensions must be equal')
+        if ndim == 2:
+            shape1 = labels[0].shape[1]
+            if not all([l.shape[1] == shape1 for l in labels]):
+                raise IOError('all labels must have same shape on 2nd dim')
 
+        # sort checking
+        for label in labels:
+            index = (np.argsort(label) if label.ndim == 1
+                     else np.lexsort(label.T))
+            # print label, index
+            # print len(index), label.shape[0]
+            assert len(index) == label.shape[0]
+            if not all(n == index[n] for n in xrange(label.shape[0]-1)):
+                raise IOError('labels are not sorted in increasing order')
+
+    @staticmethod
+    def parse_dim(labels):
+        """Return the labels vectors dimension"""
         return 1 if labels[0].ndim == 1 else labels[0].shape[1]
-        # return labels[0].ndim
 
     def __eq__(self, other):
         if self is other:
             return True
         try:
-            # check the little attributes
+            # check little attributes
             if not (self.name == other.name and
                     self.dim == other.dim and
                     self.dtype == other.dtype and